requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==2.0.0
base58==2.1.1
beautifulsoup4==4.9.3
bitarray==2.5.1
blinker==1.5
blis==0.7.9
Bottleneck==1.3.5
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.8
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
confection==0.0.3
configparser==5.2.0
cryptography==36.0.1
cycler==0.10.0
cymem==2.0.7
cytoolz==0.12.0
decorator==5.1.1
distlib==0.3.1
Django==3.1.2
django-crispy-forms==1.9.2
djangorestframework==3.12.4
entrypoints==0.4
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.4.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fastai==1.0.57
fastprogress==1.0.3
filelock==3.0.12
Flask==2.1.3
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.29
gunicorn==20.1.0
hexbytes==0.2.2
idna==2.10
importlib-metadata==4.12.0
inltk==0.9
ipfshttpclient==0.8.0a2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.7.2
kiwisolver==1.3.1
langcodes==3.3.0
lazy-object-proxy==1.7.1
lru-dict==1.1.8
lxml==4.7.1
MarkupSafe==2.1.1
matplotlib==3.4.2
mccabe==0.6.1
multiaddr==0.0.9
multidict==6.0.2
murmurhash==1.0.9
netaddr==0.8.0
nose==1.3.7
numexpr==2.8.3
numpy==1.20.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.2.4
parsimonious==0.8.1
pathy==0.6.2
Pillow==8.2.0
platformdirs==2.5.0
preshed==3.0.8
protobuf==3.20.1
pyarrow==9.0.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
pydeck==0.8.0b4
Pygments==2.13.0
pylint==2.12.2
Pympler==1.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
rich==12.6.0
rlp==2.0.1
scikit-learn==1.1.2
scipy==1.9.2
selenium==3.141.0
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.1
srsly==2.4.4
streamlit==1.13.0
thinc==8.1.4
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
tweepy==4.4.0
typer==0.4.2
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.4
validators==0.20.0
varint==1.0.2
virtualenv==20.4.6
virtualenvwrapper-win==1.2.6
wasabi==0.10.1
watchdog==2.1.9
web3==5.30.0
websockets==9.1
Werkzeug==2.1.2
whitenoise==6.4.0
wrapt==1.13.3
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.7.2
zipp==3.8.1
